requirements.txt
Adafruit-Blinka==8.69.0
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-bmp280==3.3.9
adafruit-circuitpython-busdevice==5.2.15
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-register==1.11.1
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.87.0
Adafruit-PureIO==1.1.11
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
binho-host-adapter==0.1.6
bleach==6.2.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorzero==2.0
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.61.1
fqdn==1.5.1
gpiozero==2.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
lgpio==0.2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.1.7
mistune==3.1.3
mpu6050-raspberrypi==1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
notebook_shim==0.2.4
numpy==2.3.0
opencv-python==4.11.0.86
openpyxl==3.1.5
overrides==7.7.0
packaging==25.0
pandas==3.0.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyftdi==0.57.1
Pygments==2.19.1
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pyusb==1.3.1
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
rpi-lgpio==0.6
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rpimotorlib @ git+https://github.com/gavinlyonsrepo/RpiMotorLib.git@a873181432d43d62ffb16fe9777ccea0ebf03b7b
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sysv_ipc==1.2.0
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
